restructure config to files
