+++ src/settings.ts
@@ -1,5 +1,5 @@
-import { MasterAddresses, Endpoint, Subgraph, VaultService } from "./constants/constants";
-import { ChainId } from '@usedapp/core'
+import { MasterAddresses, Endpoint, Subgraph, VaultService, Chains } from "./constants/constants";
+import { Chain, ChainId } from '@usedapp/core'
 import { NFTAirdropAddress, TokenAirdropAddress, TokenAirdropCID, TokenAirdropDelegatees } from "components/Airdrop/constants";
 
 export const CHAINID: ChainId = process.env.REACT_APP_CHAINID ? parseInt(process.env.REACT_APP_CHAINID) as ChainId : undefined || ChainId.Rinkeby;
@@ -11,6 +11,8 @@ export const NFT_AIRDROP_ADDRESS = NFTAirdropAddress[CHAINID] || NFTAirdropAddre
 export const TOKEN_AIRDROP_ADDRESS = TokenAirdropAddress[CHAINID];
 export const DELEGATEES_IPFS = process.env.REACT_APP_DELEGATEES_IPFS || TokenAirdropDelegatees[CHAINID];
 export const TOKEN_AIRDROP_IPFS_CID = process.env.REACT_APP_TOKEN_AIRDROP_IPFS_CID || TokenAirdropCID[CHAINID];
+
+export const defaultChain: Chain = Chains[CHAINID];
 
 /** TODO: temporary until we'll eliminate the polling */
 export const POLL_INTERVAL = 100000;
